Create python code
The purpose of this Project was to make a
Python coding which will enable us to
make, perform and modify the contents of
Jee_Main Applicants Record. In this
project MySQL database creation and
table creations are done by using python
code. In this project back end is MYSQL
and front end is Python.